[TASK] Simplify MailingService
This patch simplifies the MailingService by refactoring it to one class
as only HTML mails are sent by the NotificationService.
